+++ lib/services/search_location/search_location_manager.dart
@@ -1,9 +1,8 @@
-import 'package:async/async.dart';
 import 'package:trufi_core/blocs/location_search_bloc.dart';
 import 'package:trufi_core/blocs/locations/favorite_locations_cubit/favorite_locations_cubit.dart';
 import 'package:trufi_core/trufi_models.dart';
 
-abstract class RequestManager {
+abstract class SearchLocationManager {
   Future<List<TrufiPlace>> fetchLocations(
     FavoriteLocationsCubit favoriteLocationsCubit,
     LocationSearchBloc locationSearchBloc,
@@ -11,21 +10,4 @@ abstract class RequestManager {
     int limit,
     String correlationId,
   });
-
-  CancelableOperation<Plan> fetchTransitPlan(
-    TrufiLocation from,
-    TrufiLocation to,
-    String correlationId,
-  );
-
-  CancelableOperation<Plan> fetchCarPlan(
-    TrufiLocation from,
-    TrufiLocation to,
-    String correlationId,
-  );
-
-  CancelableOperation<Ad> fetchAd(
-    TrufiLocation to,
-    String correlationId,
-  );
 }
