+++ test/mocks/request_manager_mock.dart
@@ -1,4 +1,4 @@
 import 'package:mockito/mockito.dart';
-import 'package:trufi_core/repository/request_manager.dart';
+import 'package:trufi_core/services/plan_request/request_manager.dart';
 
 class MockRequestManager extends Mock implements RequestManager {}
